+++ pinot-controller/src/main/java/org/apache/pinot/controller/api/resources/services/BrokersApiServiceImpl.java
@@ -1,0 +1,4 @@
+package org.apache.pinot.controller.api.resources.services;
+
+public class BrokersApiServiceImpl extends V2ApiServiceImpl {
+}
