+++ pinot-controller/src/main/openapi/templates/apiService.mustache
@@ -1,0 +1,28 @@
+package {{package}};
+
+import {{package}}.*;
+import {{modelPackage}}.*;
+
+import org.glassfish.jersey.media.multipart.FormDataBodyPart;
+
+{{#imports}}import {{import}};
+{{/imports}}
+
+import java.util.List;
+import {{package}}.NotFoundException;
+
+import java.io.InputStream;
+
+import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
+{{#useBeanValidation}}
+import javax.validation.constraints.*;
+{{/useBeanValidation}}
+{{>generatedAnnotation}}
+{{#operations}}
+public interface {{classname}}Service {
+    {{#operation}}
+    Response {{nickname}}({{#allParams}}{{>serviceQueryParams}}{{>servicePathParams}}{{>serviceHeaderParams}}{{>serviceBodyParams}}{{>serviceFormParams}},{{/allParams}}SecurityContext securityContext) throws NotFoundException;
+    {{/operation}}
+}
+{{/operations}}
