+++ pinot-controller/src/main/openapi/templates/apiServiceFactory.mustache
@@ -1,0 +1,13 @@
+package {{package}}.factories;
+
+import {{package}}.{{classname}}Service;
+import {{package}}.services.{{classname}}ServiceImpl;
+
+{{>generatedAnnotation}}
+public class {{classname}}ServiceFactory {
+    private static final {{classname}}Service service = new {{classname}}ServiceImpl();
+
+    public static {{classname}}Service get{{classname}}() {
+        return service;
+    }
+}
